refactor: move load_state and dump_state to serializer submodule
