+++ src/autora/workflow/__main__.py
@@ -1,14 +1,13 @@
 import importlib
 import logging
 import pathlib
-from typing import Optional, Union
+from typing import Optional
 
 import typer
 from typing_extensions import Annotated
 
-from autora.state import State
 from autora.workflow.serializer import Supported as SerializersSupported
-from autora.workflow.serializer import get_serializer_mode
+from autora.workflow.serializer import _dump_state, _load_state
 
 _logger = logging.getLogger(__name__)
 
@@ -40,6 +39,7 @@ def main(
     verbose: Annotated[bool, typer.Option(help="Turns on info logging level.")] = False,
     debug: Annotated[bool, typer.Option(help="Turns on debug logging level.")] = False,
 ):
+    """Run an arbitrary function (on an optional State object) and store the output."""
     _configure_logger(debug, verbose)
     starting_state = _load_state(in_path, in_loader)
     _logger.info(f"Starting State: {starting_state}")
@@ -60,21 +60,6 @@ def _configure_logger(debug, verbose):
         _logger.info("using INFO logging level")
 
 
-def _load_state(
-    path: Optional[pathlib.Path],
-    loader: SerializersSupported = SerializersSupported.dill,
-) -> Union[State, None]:
-    if path is not None:
-        load, file_mode = get_serializer_mode(loader, "load")
-        _logger.debug(f"_load_state: loading from {path=}")
-        with open(path, f"r{file_mode}") as f:
-            state_ = load(f)
-    else:
-        _logger.debug(f"_load_state: {path=} -> returning None")
-        state_ = None
-    return state_
-
-
 def _load_function(fully_qualified_function_name: str):
     _logger.debug(f"_load_function: Loading function {fully_qualified_function_name}")
     module_name, function_name = fully_qualified_function_name.rsplit(".", 1)
@@ -82,24 +67,6 @@ def _load_function(fully_qualified_function_name: str):
     function = getattr(module, function_name)
     _logger.debug(f"_load_function: Loaded function {function} from {module}")
     return function
-
-
-def _dump_state(
-    state_: State,
-    path: Optional[pathlib.Path],
-    dumper: SerializersSupported = SerializersSupported.dill,
-) -> None:
-    if path is not None:
-        dump, file_mode = get_serializer_mode(dumper, "dump")
-        _logger.debug(f"_dump_state: dumping to {path=}")
-        path.parent.mkdir(parents=True, exist_ok=True)
-        with open(path, f"w{file_mode}") as f:
-            dump(state_, f)
-    else:
-        dumps, _ = get_serializer_mode(dumper, "dumps")
-        _logger.debug(f"_dump_state: {path=} so writing to stdout")
-        print(dumps(state_))
-    return
 
 
 if __name__ == "__main__":
